Split automate_diablo into helpers and main
